Move src/contextprops into src/preact
