+++ extensions/amp-accordion/1.0/test/test-amp-accordion.js
@@ -16,7 +16,7 @@
 import '../amp-accordion';
 import {ActionInvocation} from '../../../../src/service/action-impl';
 import {ActionTrust} from '../../../../src/action-constants';
-import {CanRender} from '../../../../src/contextprops';
+import {CanRender} from '../../../../src/preact/contextprops';
 import {htmlFor} from '../../../../src/static-template';
 import {subscribe, unsubscribe} from '../../../../src/context';
 import {toggleExperiment} from '../../../../src/experiments';
